[REFACTOR] Move client scss files into folder
