+++ src/client/react/admin/views/ChallengesView/ChallengeProfile.js
@@ -17,7 +17,7 @@ import {
 import NotificationToaster from '../../components/NotificationToaster';
 import MarkdownEditor from '../../../../../../lib/react/components/MarkdownEditor';
 
-import '../../../scss/components/_instruction-panel.scss';
+import '../../../user/scss/components/_instruction-panel.scss';
 import '../../scss/_markdown-preview.scss';
