+++ src/client/react/user/components/Description.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ScrollAnimation from 'react-animate-on-scroll';
 
-import '../../scss/components/_description.scss';
+import '../scss/components/_description.scss';
 
 
 class Description extends React.Component {
