+++ src/client/react/user/components/Title.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ScrollAnimation from 'react-animate-on-scroll';
 
-import '../../scss/components/_hero.scss';
+import '../scss/components/_hero.scss';
 import '../../../assets/images/hero-desktop-min.png';
 import '../../../assets/images/hero-mobile-min.png';
