+++ src/client/react/user/scss/components/_base.scss
@@ -1,4 +1,4 @@
-@import '../variables/bp';
+@import '../../../scss/variables/bp';
 
 .hero {
 	#base {
