+++ src/client/react/user/scss/components/_dashboard.scss
@@ -1,5 +1,5 @@
-@import '../variables/bp';
-@import '../variables/colours';
+@import '../../../scss/variables/bp';
+@import '../../../scss/variables/colours';
 
 #dashboard {
 	.dashboard {
