+++ src/client/react/user/scss/components/_help-menu.scss
@@ -1,4 +1,4 @@
-@import '../variables/colours';
+@import '../../../scss/variables/colours';
 
 #help-menu {
 	.help-button {
