+++ src/client/react/user/scss/components/_hero.scss
@@ -1,6 +1,6 @@
-@import '../variables/bp';
-@import '../variables/colours';
-@import '../variables/fonts';
+@import '../../../scss/variables/bp';
+@import '../../../scss/variables/colours';
+@import '../../../scss/variables/fonts';
 
 $left-margin: 3rem;
 $left-margin-s: 1rem;
