+++ src/client/react/user/scss/components/_instruction-panel.scss
@@ -1,4 +1,4 @@
-@import '../../scss/variables/colours';
+@import '../../../scss/variables/colours';
 
 .instruction-panel {
 	.markdown.output.markdown-content, .markdown.output.help {
