+++ src/client/react/user/scss/components/_pay.scss
@@ -1,6 +1,5 @@
-@import '../variables/bp';
-@import '../variables/colours';
-
+@import '../../../scss/variables/bp';
+@import '../../../scss/variables/colours';
 
 #pay {
 	.infobox {
