+++ src/client/react/user/scss/components/_registration-closed.scss
