+++ src/client/react/user/scss/views/_main.scss
@@ -1,4 +1,4 @@
-@import '../variables/colours';
+@import '../../../scss/variables/colours';
 
 .dashboard {
 	max-width: 40rem;
