+++ src/client/react/user/scss/views/_team-panel.scss
