+++ src/client/react/user/stages/RegistrationClosed/Home.js
@@ -7,7 +7,7 @@ import Title from '../../components/Title';
 import Description2 from '../../components/Description2';
 import ScrollAnimation from 'react-animate-on-scroll';
 
-import '../../../scss/components/_registration-closed.scss';
+import '../../scss/components/_registration-closed.scss';
 
 
 const mapStateToProps = (state, ownProps) => {
