+++ src/client/react/user/views/Instructions/InstructionCollapse.js
@@ -2,8 +2,8 @@ import React from 'react';
 import autobind from 'core-decorators/es/autobind';
 import { Button, Intent, Collapse } from '@blueprintjs/core';
 import MarkdownRenderer from '../../../../../../lib/react/components/MarkdownRenderer';
-import '../../../scss/components/_instruction-collapse.scss'
-import '../../../scss/components/_instruction-panel.scss';
+
+import '../../scss/components/_instruction-panel.scss';
 
 
 @autobind
